Refactor out SingleResponseCollector + interface
