+++ rpc/response/ResponseCollector.java
@@ -19,12 +19,44 @@
 
 package grakn.client.rpc.response;
 
-public interface ResponseCollector {
-    /**
-     * Callback function called when a response is received. The return value determines whether this ResponseCollector
-     * is done and should be cleaned up.
-     * @param response The response.
-     * @return 'true' if this ResponseCollector is done receiving responses, otherwise, 'false'.
-     */
-    boolean onResponse(Response response);
+import grakn.protocol.TransactionProto;
+
+import java.util.concurrent.BlockingQueue;
+import java.util.concurrent.LinkedBlockingQueue;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
+
+public class ResponseCollector {
+    private volatile boolean isDone;
+    private final BlockingQueue<Response> received = new LinkedBlockingQueue<>();
+
+    public ResponseCollector() {
+        isDone = false;
+    }
+
+    public void onResponse(final Response response) {
+        received.add(response);
+        TransactionProto.Transaction.Res res = response.response();
+        if (res == null || isLastResponse(res)) {
+            isDone = true;
+        }
+    }
+
+    public boolean isDone() {
+        return isDone;
+    }
+
+    public TransactionProto.Transaction.Res take() throws InterruptedException {
+        return received.take().ok();
+    }
+
+    public TransactionProto.Transaction.Res take(long timeout, TimeUnit unit) throws InterruptedException {
+        return received.poll(timeout, unit).ok();
+    }
+
+    boolean isLastResponse(final TransactionProto.Transaction.Res response) {
+        if (!response.hasIterRes()) return true;
+        final TransactionProto.Transaction.Iter.Res iterRes = response.getIterRes();
+        return iterRes.getIteratorID() != 0 || iterRes.getDone();
+    }
 }
